# VRTDerivedRaster: optimizations, in particular for sum(), and make 'gdal raster calc' use sum() when possible