--- a/apps/gdalalg_raster_calc.cpp
+++ b/apps/gdalalg_raster_calc.cpp
@@ -227,6 +227,39 @@ UpdateSourceProperties(SourceProperties &out, const std::string &dsn,
     return source;
 }
 
+static bool IsSumAllSources(const std::string &expression,
+                            const std::map<std::string, std::string> &sources)
+{
+    bool ok = false;
+    std::string modExpression;
+    const char *pszSep = "+ ";
+    if (cpl::starts_with(expression, "sum(") && expression.back() == ')')
+    {
+        modExpression = expression.substr(4, expression.size() - 5);
+        pszSep = ", ";
+    }
+    else
+    {
+        modExpression = expression;
+    }
+    const CPLStringList aosTokens(
+        CSLTokenizeString2(modExpression.c_str(), pszSep, 0));
+    if (static_cast<size_t>(aosTokens.size()) == sources.size())
+    {
+        std::set<std::string> foundSources;
+        for (int i = 0; i < aosTokens.size(); ++i)
+        {
+            if (cpl::contains(sources, aosTokens[i]) &&
+                !cpl::contains(foundSources, aosTokens[i]))
+            {
+                foundSources.insert(aosTokens[i]);
+            }
+        }
+        ok = foundSources.size() == sources.size();
+    }
+    return ok;
+}
+
 /** Create XML nodes for one or more derived bands resulting from the evaluation
  *  of a single expression
  *
@@ -246,6 +279,8 @@ CreateDerivedBandXML(CPLXMLNode *root, int nXOut, int nYOut,
                      const std::map<std::string, SourceProperties> &sourceProps,
                      const std::string &fakeSourceFilename)
 {
+    const bool bIsSumAllSources = IsSumAllSources(expression, sources);
+
     int nOutBands = 1;  // By default, each expression produces a single output
                         // band. When processing the expression below, we may
                         // discover that the expression produces multiple bands,
@@ -261,12 +296,6 @@ CreateDerivedBandXML(CPLXMLNode *root, int nXOut, int nYOut,
         CPLAddXMLAttributeAndValue(band, "subClass", "VRTDerivedRasterBand");
         CPLAddXMLAttributeAndValue(band, "dataType",
                                    GDALGetDataTypeName(bandType));
-
-        CPLXMLNode *pixelFunctionType =
-            CPLCreateXMLNode(band, CXT_Element, "PixelFunctionType");
-        CPLCreateXMLNode(pixelFunctionType, CXT_Text, "expression");
-        CPLXMLNode *arguments =
-            CPLCreateXMLNode(band, CXT_Element, "PixelFunctionArguments");
 
         for (const auto &[source_name, dsn] : sources)
         {
@@ -363,11 +392,24 @@ CreateDerivedBandXML(CPLXMLNode *root, int nXOut, int nYOut,
             }
         }
 
-        // Add the expression as a last step, because we may modify the
-        // expression as we iterate through the bands.
-        CPLAddXMLAttributeAndValue(arguments, "expression",
-                                   bandExpression.c_str());
-        CPLAddXMLAttributeAndValue(arguments, "dialect", "muparser");
+        CPLXMLNode *pixelFunctionType =
+            CPLCreateXMLNode(band, CXT_Element, "PixelFunctionType");
+        if (bIsSumAllSources)
+        {
+            CPLCreateXMLNode(pixelFunctionType, CXT_Text, "sum");
+        }
+        else
+        {
+            CPLCreateXMLNode(pixelFunctionType, CXT_Text, "expression");
+            CPLXMLNode *arguments =
+                CPLCreateXMLNode(band, CXT_Element, "PixelFunctionArguments");
+
+            // Add the expression as a last step, because we may modify the
+            // expression as we iterate through the bands.
+            CPLAddXMLAttributeAndValue(arguments, "expression",
+                                       bandExpression.c_str());
+            CPLAddXMLAttributeAndValue(arguments, "dialect", "muparser");
+        }
     }
 
     return true;