--- a/gcore/gdal_misc.cpp
+++ b/gcore/gdal_misc.cpp
@@ -68,6 +68,11 @@ static int GetMinBitsForPair(const bool pabSigned[], const bool pabFloating[],
 
     if (pabSigned[0] != pabSigned[1])
     {
+        if (!pabSigned[0] && panBits[0] < panBits[1])
+            return panBits[1];
+        if (!pabSigned[1] && panBits[1] < panBits[0])
+            return panBits[0];
+
         const int nUnsignedTypeIndex = pabSigned[0] ? 1 : 0;
         const int nSignedTypeIndex = pabSigned[0] ? 0 : 1;
 
@@ -78,7 +83,7 @@ static int GetMinBitsForPair(const bool pabSigned[], const bool pabFloating[],
     return std::max(panBits[0], panBits[1]);
 }
 
-static int GetDataTypeElementSizeBits(GDALDataType eDataType)
+static int GetNonComplexDataTypeElementSizeBits(GDALDataType eDataType)
 {
     switch (eDataType)
     {
@@ -136,8 +141,8 @@ GDALDataType CPL_STDCALL GDALDataTypeUnion(GDALDataType eType1,
     if (eType2 == GDT_Unknown)
         return eType1;
 
-    const int panBits[] = {GetDataTypeElementSizeBits(eType1),
-                           GetDataTypeElementSizeBits(eType2)};
+    const int panBits[] = {GetNonComplexDataTypeElementSizeBits(eType1),
+                           GetNonComplexDataTypeElementSizeBits(eType2)};
 
     if (panBits[0] == 0 || panBits[1] == 0)
         return GDT_Unknown;
@@ -149,12 +154,11 @@ GDALDataType CPL_STDCALL GDALDataTypeUnion(GDALDataType eType1,
     const bool pabFloating[] = {CPL_TO_BOOL(GDALDataTypeIsFloating(eType1)),
                                 CPL_TO_BOOL(GDALDataTypeIsFloating(eType2))};
     const bool bFloating = pabFloating[0] || pabFloating[1];
-    const bool bComplex = CPL_TO_BOOL(GDALDataTypeIsComplex(eType1)) ||
-                          CPL_TO_BOOL(GDALDataTypeIsComplex(eType2));
-
     const int nBits = GetMinBitsForPair(pabSigned, pabFloating, panBits);
+    const bool bIsComplex = CPL_TO_BOOL(GDALDataTypeIsComplex(eType1)) ||
+                            CPL_TO_BOOL(GDALDataTypeIsComplex(eType2));
 
-    return GDALFindDataType(nBits, bSigned, bFloating, bComplex);
+    return GDALFindDataType(nBits, bSigned, bFloating, bIsComplex);
 }
 
 /************************************************************************/