… gcore/gdalsse_priv.h