- GDALDataTypeUnion(): adjust to avoid Byte + CInt32 to go to CFloat64
- VRTDerivedRasterBand::IRasterIO(): select a smaller source transfer type that the output buffer type if possible
- gdal raster calc: remove setting SourceTransferType (no longer needed since above commit)
- VRTDerivedRasterBand::IRasterIO(): do not zero/nodata initialize source buffer when not needed
-  VRT pixel function: optimize 'sum' for common data types
    
    Summing two 20k x 20k Byte rasters,
    
    with this PR:
    ```
    |+--------------+--------------------+--------+
    | Transfer type | VRTRasterBand type | timing |
    |+--------------+------------------- + -------+
    | Int16/UInt16  | Float32            | 1.4s   |
    | Int32         | Float32            | 1.9s   |
    | Int32         | Int32              | 2.5s   |
    | Byte          | Float32            | 1.2s   |
    | Byte          | Float64            | 2.1s   |
    | UInt32        | UInt32             | 11.8s  |  (no optimized template specialization for this likely rarely used type)
    |+--------------+--------------------+--------+
    ```
    with master:
    ```
    |+--------------+--------------------+--------+
    | Transfer type | VRTRasterBand type | timing |
    |+--------------+------------------- + -------+
    | Int16/UInt16  | Float32            | 12.4s  |
    | Int32         | Float32            | 12.8s  |
    | Int32         | Int32              | 13.2s  |
    | Byte          | Float32            | 11.8s  |
    | Byte          | Float64            | 11.4s  |
    | UInt32        | UInt32             | 13.0s  |
    |+--------------+--------------------+--------+
    ```
- gdal raster calc: detect providing twice the same input name
- gdal raster calc: check input names are valid identifiers
- VRT pixelfunction sum(): accept one single input
- gdal raster calc: optimize sum of all sources as using builtin sum()

CC @dbaston