@dbaston 2025-05-18
I had been looking into this as well. I saw noticeable but surprisingly small difference between muparser and the built-in "sum" function (pre-dating these optimizations, of course). But I'm not sure how common summing integer rasters is.

1000x1000, 2 bands
```
dan@t14:~/dev/gdal/cmake-build-release$ python3 ~/dev/gdal-scratch/pixfn_bench.py 1000 1000 2 expression '{"expression":"sum(BANDS)"}'
230 executions, median time: 0.015510411991272122
dan@t14:~/dev/gdal/cmake-build-release$ python3 ~/dev/gdal-scratch/pixfn_bench.py 1000 1000 2 sum ""
392 executions, median time: 0.011796439997851849
```

1000x1000, 20 bands
```
dan@t14:~/dev/gdal/cmake-build-release$ python3 ~/dev/gdal-scratch/pixfn_bench.py 1000 1000 20 sum ""
91 executions, median time: 0.07296971406321973
dan@t14:~/dev/gdal/cmake-build-release$ python3 ~/dev/gdal-scratch/pixfn_bench.py 1000 1000 20 expression '{"expression":"sum(BANDS)"}'
87 executions, median time: 0.08568526105955243
```