@rouault 2025-05-18
> I saw noticeable but surprisingly small difference between muparser

I'm not that surprised. There is necessarily an overhead going through muparser which must have some sort of interpreter rather than the wired builtin sum()